Collect the attribute vocabulary in one file
Naming what sshmux observes is signal-agnostic, whichever signal happens to
consume it: outcomeAttributes resolves an error the same way whether a counter
or a span carries it, and serverAttributes and peerAttributes only ever read
the name table their receiver held.

Move the names, the conventions and every builder over them into attributes.go,
each a method on attributeNames, and have the builders drop what a convention
has no name for so that no caller has to. The tracer keeps one-line bindings
for the three whose callers are in sshmux.go, and the metrics keep
connectionAttributeSet, which is their cardinality policy rather than a naming
decision. TestConventionAttributeNames and TestErrorType follow the code they
cover.

Two comments went with it that no longer described anything: attributeNames'
was orphaned above connectionInfo's, and the ECS table's denied naming anything
apart from the default one, which the comment ten lines below it contradicts.
